refactor(api): isolate vendor creation handler
