+++ tools/archtests/bootstrap_surface_test.go
@@ -176,10 +176,10 @@ import (
 // tenant authorization, dependency wiring, response mapping, and cache
 // invalidation. Resolver mapping, receipt verification, and closure decisions
 // remain in internal/securitylifecyclefindings and internal/findings.
-// Manual GRC vendor creation adds bounded request decoding, resolved-scope
-// authorization, event-envelope mapping, append/projection wiring, and response
-// mapping. Portable vendor state and read behavior remain in internal/grcvendor.
-const bootstrapProductionGoLineBudget = 30143
+// Manual GRC vendor creation leaves only scope/auth adapters, dependency wiring,
+// cache invalidation, and route handoff in bootstrap. Bounded request decoding,
+// event mapping, and append/projection ordering stay in sourcehttp/grcvendorcreate.
+const bootstrapProductionGoLineBudget = 29817
 
 type bootstrapFileLineCount struct {
 	path  string
